+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 This log documents significant changes for each release.  This project follows
 [Semantic Versioning](http://semver.org/).
 
+## [4.5.1] - 2025-07-09
+### Changed
+- Updated the "%terminologies.validateVS()" method and "memberOf()" function to
+  use GET for CodeableConcept with a single Coding when using a ValueSet
+  reference instead of an actual ValueSet.
+
 ## [4.5.0] - 2025-07-01
 ### Added
 - additional function `resolve()`.